myBatis-Plus集成

+++ rosycloud-common/src/main/java/cn/rosycloud/config/Constants.java
@@ -23,9 +23,9 @@ public class Constants {
     /**
      * 人员类型
      */
-    public static final Short User_Normal=1;//正常
-    public static final Short User_Forbidden=0;//禁用
-    public static final Short User_ADMIN=-1;//超级管理员
+    public static final Integer User_Normal=1;//正常
+    public static final Integer User_Forbidden=0;//禁用
+    public static final Integer User_ADMIN=-1;//超级管理员
 
     /**
      * 用户类型：1、系统类型用户
@@ -44,28 +44,28 @@ public class Constants {
     /**
      * 删除
      */
-    public static final Short Delete_Forbidden=1;
+    public static final Integer Delete_Forbidden=1;
     /**
      * 正常
      */
-    public static final Short Delete_Normal=0;
+    public static final Integer Delete_Normal=0;
 
     /**
      *日志级别定义
      */
-    public static final Short Log_Leavel_INFO=1;
-    public static final Short Log_Leavel_WARRING=2;
-    public static final Short Log_Leavel_ERROR=3;
+    public static final Integer Log_Leavel_INFO=1;
+    public static final Integer Log_Leavel_WARRING=2;
+    public static final Integer Log_Leavel_ERROR=3;
     /**
      * 日志类型
      */
-    public static final Short Log_Type_LOGIN=1; //登陆
-    public static final Short Log_Type_EXIT=2;  //退出
-    public static final Short Log_Type_INSERT=3; //插入
-    public static final Short Log_Type_DEL=4; //删除
-    public static final Short Log_Type_UPDATE=5; //更新
-    public static final Short Log_Type_UPLOAD=6; //上传
-    public static final Short Log_Type_OTHER=7; //其他
+    public static final Integer Log_Type_LOGIN=1; //登陆
+    public static final Integer Log_Type_EXIT=2;  //退出
+    public static final Integer Log_Type_INSERT=3; //插入
+    public static final Integer Log_Type_DEL=4; //删除
+    public static final Integer Log_Type_UPDATE=5; //更新
+    public static final Integer Log_Type_UPLOAD=6; //上传
+    public static final Integer Log_Type_OTHER=7; //其他
 
 
 }
